kafka-consumer/requirements.txt
kafka-python==2.0.2
mysql-connector-python==8.0.33
